+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Reorganize the folder structure and change CS standard (#405)
 - [BC BREAK] Remove the `monolog` configuration option. Instead, register the service manually (#406)
 - [BC BREAK] Remove the `listener_priorities` configuration option. Instead, use a compiler pass to change the priority of the listeners (#407)
+- Prefer usage of the existing `Symfony\Bridge\PsrHttpMessage\HttpMessageFactoryInterface` service for the `RequestFetcher` class (#409)
 
 ## 3.5.3 (2020-10-13)